requirements.txt
flask
flask-sqlalchemy
flask-migrate
fpdf2
num2words
openpyxl
pandas
python-dotenv
gunicorn
